+++ kernel/lib/actionCreators/index.js
@@ -10,6 +10,7 @@ module.exports = {
   transaction: require('./transaction'),
   update: require('./update'),
   utils: require('./utils'),
-  subscription: require('./subscription'),
-  seed: require('./seed')
+  subscription: require('./wallet'),
+  seed: require('./seed'),
+  sendAra: require('./sendAra')
 }