+++ kernel/lib/actions/afsManager.js
@@ -1,6 +1,7 @@
 'use strict'
 
 const debug = require('debug')('acm:kernel:lib:actions:afsManager')
+const k = require('../../../lib/constants/stateManagement')
 const actionsUtil = require('./utils')
 const fs = require('fs')
 const araFilesystem = require('ara-filesystem')
@@ -40,7 +41,7 @@ async function exportFolder({ did, exportPath, folderPath, completeHandler }) {
     }, { name: exportPath }, { keepExisting: true })
     await new Promise((accept, reject) => progress.once('end', accept).once('error', reject))
     await afs.close()
-    shell.openItem(path.dirname(exportPath))
+    shell.openItem(exportPath)
     completeHandler()
     debug('Successfully exported folder')
   } catch (err) {
@@ -97,6 +98,17 @@ async function _getContentsInFolder(afs, folderPath) {
   }
 }
 
+async function getUpdateAvailableStatus(item) {
+  let updateAvailable
+  try {
+    updateAvailable = await araFilesystem.isUpdateAvailable({ did: item.did })
+  } catch (err) {
+    debug('Error getting update available status')
+  }
+  return { ...item, status: updateAvailable && item.downloadPercent === 1 ? k.UPDATE_AVAILABLE : item.status }
+}
+
+
 async function surfaceAFS({ dids, DCDNStore, published = false }) {
 	return Promise.all(dids.map(did => actionsUtil.descriptorGenerator(did, {
 		shouldBroadcast: farmerManager.getBroadcastingState({ did, DCDNStore }),
@@ -116,7 +128,8 @@ function unarchiveAFS({ did }) {
 module.exports = {
   exportFile,
   exportFolder,
-	getFileList,
+  getFileList,
+  getUpdateAvailableStatus,
 	surfaceAFS,
 	unarchiveAFS,
 }